Min-Max Sum | HackerRank
Given five positive integers, find the minimum and maximum values that can be calculated by summing exactly four of the five integers. Then print the respective minimum and maximum values as a single line of two space-separated long integers.
For example,arr = [1,3,5,7,9] . Our minimum sum is 1 + 3 + 5 + 7 and our maximum sum is  3 + 5 + 7 + 9. We would print
16 24
Function Description :
Complete the miniMaxSum function in the editor below. It should print two space-separated integers on one line: the minimum sum and the maximum sum of 4 of 5 elements.
miniMaxSum has the following parameter(s):
arr: an array of 5 integers